// management/frontend/src/api/schemas/deviceBindingPublic.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Ego Collection Management
 * OpenAPI spec version: 0.1.0
 */

export interface DeviceBindingPublic {
  id?: string;
  created_at?: string | null;
  updated_at?: string | null;
  deleted_at?: string | null;
  creator_id?: string | null;
  updater_id?: string | null;
  deleter_id?: string | null;
  /**
     * @minLength 1
     * @maxLength 128
     */
  serial_number: string;
  /** @maxLength 128 */
  pid?: string;
  /**
     * @minLength 1
     * @maxLength 128
     */
  device_name: string;
  /** @maxLength 128 */
  device_model?: string;
  /** @maxLength 64 */
  slot_role?: string;
  physical_kit_id?: string | null;
  /** @maxLength 32 */
  status?: string;
  /** @maxLength 64 */
  firmware_version?: string;
  last_seen_at?: string | null;
  /** @maxLength 1000 */
  remark?: string;
  workspace_id: string;
  creator_name?: string | null;
  updater_name?: string | null;
}